requirements.txt
alembic==1.4.2
dnspython==2.1.0
Flask==1.1.2
Flask_Login==0.5.0
Flask_Migrate==2.5.3
Flask_PyMongo==2.3.0
Flask_SQLAlchemy==2.4.4
Flask_WTF==0.14.3
pymongo==3.11.3
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.1

